Unifies Message and RawMessage
Since the introduction of metadata to goduck, the Message struct and the
RawMessage interface became so similar, that it wasn't necessary to have both
anymore.

The idea behind the RawMessage interface is that the Stream/MessagePool could
attach specific metadata, so that it could correctly handle acks/nacks. This
can still be achieved with only one interface.

This commit also removes all the unit tests, since they were complex and hard
to maintain. They will be rewritten before the 1.0 version.